+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added options to use the SLP address instead of providing `--host`.
     - `--slp-host-imsi` to use the IMSI to generate the SLP address.
     - `--slp-host-cell` to use the cell information to generate the SLP address.
+- Fixed a bug where the default `ura-override` value was initialized to 0 causing the SPARTN generator to always include it as "unknown".
 
 ## [3.4.9] 2024-05-03
 - Added a few options to controll how SPARTN messages are generated: